reafactor creating sub compontens
