+++ frontend/src/Pages/RecruitmentApplicationsOverviewPage/RecruitmentApplicationsOverviewPage.module.scss
@@ -28,42 +28,3 @@ $back-button-width: 5rem;
 .empty_div {
   width: $back-button-width;
 }
-
-.arrows {
-  &:hover {
-    filter: brightness(150%);
-    transform: scale(1.05);
-  }
-  &:active {
-    filter: brightness(200%);
-    transform: scale(1.10);
-  }
-}
-
-.withdrawnHeader {
-  background-color: $red-samf;
-  color: $white;
-  &:hover {
-    background-color: $red-samf;
-    filter: brightness(95%);
-  }
-  border:none;
-}
-
-.withdrawnRow {
-  background-color: $grey-3;
-  &:hover {
-    background-color: $grey-3;
-    filter: brightness(95%);
-  }
-  border-bottom: none;
-  border-top: 1px solid $grey-2;
-}
-
-.withdrawnContainer {
-  margin-top: 2em;
-}
-
-.withdrawnText {
-  color: $red;
-}
