+++ frontend/src/Pages/RecruitmentApplicationsOverviewPage/components/ActiveApplications/index.ts
@@ -1,0 +1,1 @@
+export { ActiveApplications } from './ActiveApplications';
