+++ frontend/src/Pages/RecruitmentApplicationsOverviewPage/components/WithdrawnApplications/WithdrawnApplications.module.scss
@@ -1,0 +1,31 @@
+@import 'src/mixins';
+
+@import 'src/constants';
+
+.withdrawnHeader {
+  background-color: $red-samf;
+  color: $white;
+  &:hover {
+    background-color: $red-samf;
+    filter: brightness(95%);
+  }
+  border: none;
+}
+
+.withdrawnRow {
+  background-color: $grey-3;
+  &:hover {
+    background-color: $grey-3;
+    filter: brightness(95%);
+  }
+  border-bottom: none;
+  border-top: 1px solid $grey-2;
+}
+
+.withdrawnContainer {
+  margin-top: 2em;
+}
+
+.withdrawnText {
+  color: $red;
+}
