+++ frontend/src/Pages/RecruitmentApplicationsOverviewPage/components/WithdrawnApplications/index.ts
@@ -1,0 +1,1 @@
+export { WithdrawnApplications } from './WithdrawnApplications';
